rework project structure
