+++ models/databaseConfig.go
@@ -1,0 +1,3 @@
+package models
+
+// TODO: Add database config structure
